feat(order): 实现订单管理功能
- 新增订单控制器，支持教师和学生查询订单列表及详情
- 实现教师为预约创建订单的功能，限制仅当预约状态为PENDING时可创建
- 支持学生支付订单和教师取消订单的操作
- 添加订单实体类字段类型调整，将price由Long改为BigDecimal以支持精确金额计算
- 更新UserOrderMapper接口与XML文件，优化查询逻辑并增加按教师ID和用户ID筛选订单的方法
- 引入定时任务自动取消超时未支付的订单（每分钟检查一次）
- 增加OrderStateConvert转换器，用于处理请求参数中的订单状态枚举值
- 完善订单相关DTO和Result对象，确保数据传输结构清晰且完整

+++ src/main/java/com/yoru/qingxintutor/pojo/entity/UserOrderEntity.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -17,8 +18,8 @@ public class UserOrderEntity {
     private Long reservationId;     // BIGINT NOT NULL
     private String item;            // VARCHAR(100) NOT NULL
     private Integer quantity;       // INT DEFAULT 1
-    private Long price;             // BIGINT NOT NULL, 单位：分
-    private State state;           // VARCHAR(20) NOT NULL 
+    private BigDecimal price;       // decimal(10,2) NOT NULL
+    private State state;            // enum NOT NULL 
     private LocalDateTime createTime;
 
     public enum State {
